chore: clean up includes, rename files
