+++ include/Commands.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "ShellClass.hpp"
+#include "Shell.hpp"
 
 #include <string>
 #include <unordered_map>
