+++ src/commands/lua.cpp
@@ -1,7 +1,5 @@
 #include "CliPrompt.hpp"
 #include "Commands.hpp"
-#include "Consoles.hpp"
-// #include "Shell.hpp"
 
 #include <sol/sol.hpp>
