+++ src/commands/wifi.cpp
@@ -1,5 +1,4 @@
 #include "Commands.hpp"
-#include "Consoles.hpp"
 
 #include <dswifi9.h>
 #include <nds.h>
