+++ src/main.cpp
@@ -1,5 +1,5 @@
 #include "Consoles.hpp"
-#include "ShellClass.hpp"
+#include "Shell.hpp"
 
 #include <algorithm>
 #include <dswifi9.h>
